feature: implement the app layout in case of loading or connection error
Define the function BookshelfLoadingScreen to implement the application layout in case there is a data loading from network that is currently running.
Define the function BookshelfErrorScreen to implement the application layout in case there is a connection error.
Save the necessary resources inside the resources folder: strings.xml, dimens.xml, drawable folder.

Integrates #1