feature: implement the app layout in case of successful data retrieval
Define the function BookshelfSuccessScreen to implement the application layout in case data are loaded correctly, to display on the screen as expected.
Import inside the build.gradle.kts file the necessary dependencies (retrofit, serialization and coil) in order to convert the image from url to drawable.
Insert the necessary dimension resources in the dimens.xml file.
Define the data class BookshelfItem that will be used as the one containing the elements constituting the UI layer.

Integrates #1