feature: define data classes for json parsing and retrofit API
In the network package, include the API definition of the web service that we are going to call.
Define then in the model package all the classes that are needed for parsing the json payload retrieved from the web service. Include only the necessary elements, the ones that will be used in the application layout.

Integrates #3