feature: define the ui state class and implement the viewModel
Define the BookshelfUiState that is going to be used as the UI state associated to the view model. Define also the BookshelfNetworkState class to differentiate the output according to the result of the REST calls.
Let the MainAppContainer be an object cause only one instance must be present at runtime.
Implement the view model logic and in the companion object define the view model factory to instantiate the view model.

Integrates #5